Move CLI arbitrage commands to a separate CLI
