+++ src/cli/arbitrageCommands/arbClaimBuyerFundsCmd.js
@@ -1,8 +1,8 @@
-const cliUtils = require('../../helpers/cliUtils')
+const cliUtils = require('../helpers/cliUtils')
 
-const getAddress = require('../../../helpers/getAddress')
-const getArbitrageContractAddress = require('../../helpers/getArbitrageContractAddress')
-const getArbitrageService = require('../../../services/ArbitrageService')
+const getAddress = require('../../helpers/getAddress')
+const getArbitrageContractAddress = require('../helpers/getArbitrageContractAddress')
+const getArbitrageService = require('../../services/ArbitrageService')
 
 function registerCommand ({ cli, logger }) {
   cli.command(
