+++ src/cli/arbitrageCommands/arbTransferEtherCmd.js
@@ -1,9 +1,9 @@
-const cliUtils = require('../../helpers/cliUtils')
-const { toWei, fromWei } = require('../../../helpers/numberUtil')
+const cliUtils = require('../helpers/cliUtils')
+const { toWei, fromWei } = require('../../helpers/numberUtil')
 
-const getAddress = require('../../../helpers/getAddress')
-const getArbitrageContractAddress = require('../../helpers/getArbitrageContractAddress')
-const getArbitrageService = require('../../../services/ArbitrageService')
+const getAddress = require('../../helpers/getAddress')
+const getArbitrageContractAddress = require('../helpers/getArbitrageContractAddress')
+const getArbitrageService = require('../../services/ArbitrageService')
 
 function registerCommand ({ cli, logger }) {
   cli.command(
